Images in Dashboard implemented The Like changes according to if it is already liked or not in the list Split MyPHP stuff up
